[Rename 🏷️] 내 운동 페이지 이름 변경 (MypageHome → MyPage)
