+++ src/routes/Navigator.tsx
@@ -5,7 +5,7 @@ import MainLayout from "@Layout/Mainlayout"
 import Detail from "@pages/Detail/Detail"
 import Home from "@pages/Home/Home"
 import Login from "@pages/Login/Login"
-import Mypage from "@pages/MyPage/My/MypageHome"
+import MyPage from "@pages/MyPage/My/MyPage"
 import BodyPart from "@pages/Recommend/BodyPart/BodyPart"
 import Machine from "@pages/Recommend/Machine/Machine"
 import Result from "@pages/Recommend/Result/Result"
@@ -55,7 +55,7 @@ const router = createBrowserRouter([
       },
       {
         path: "mypage",
-        element: <Mypage />,
+        element: <MyPage />,
       },
     ],
   },
